backend/src/requirements.txt
djangorestframework==3.11.0

#asgiref==3.2.10
#django==3.0.8
#pytz==2020.1
#sqlparse==0.3

